requirements.txt
pytest==7.4.0
selene==2.0.0rc4
Faker==18.11.1
selenium
allure-pytest==2.13.2
python-dotenv==0.20.0